requirements.txt
# requirements.txt

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Your existing dependencies (keep these exact versions)
langchain==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38
openai==1.51.2
pinecone-client==3.2.2
httpx>=0.25.0,<0.28.0
python-dotenv==1.0.1
numpy==1.26.4

# PDF processing (your existing)
PyPDF2==3.0.1
pdfplumber==0.10.3

# Optional: Production server
gunicorn==21.2.0

# Enhanced dependencies for improved functionality
# Performance and monitoring
psutil>=5.9.0
prometheus-client>=0.17.0

# Enhanced error handling and retry logic
tenacity>=8.2.0
backoff>=2.2.0

# Advanced data validation
marshmallow>=3.20.0
cerberus>=1.3.0

# Enhanced logging and monitoring
structlog>=23.1.0
colorama>=0.4.6

# Performance profiling
py-spy>=0.3.14
memory-profiler>=0.61.0

# Security enhancements
cryptography>=41.0.0
passlib>=1.7.4

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality and formatting
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Optional: Advanced features
# Redis for caching (optional)
redis==5.0.4
fastapi-cache2==0.2.1

# Celery for background tasks (optional)
celery>=5.3.0

# Advanced monitoring (optional)
sentry-sdk>=1.28.0
